refactor(face-cleanup): remove the retired 2-state apply endpoint
