+++ server/src/dtos/face-repair.dto.ts
@@ -123,31 +123,6 @@ export const FaceRepairScanStatusSchema = z
   .meta({ id: 'FaceRepairScanStatusDto' });
 export class FaceRepairScanStatusDto extends createZodDto(FaceRepairScanStatusSchema) {}
 
-const FaceRepairManualMoveSchema = z.object({
-  personId: z.uuidv4(),
-  destinationPersonId: z.uuidv4(),
-  faceIds: z.array(z.uuidv4()).optional(),
-  entireCluster: z.boolean().optional(),
-});
-
-export const FaceRepairApplyRequestSchema = z
-  .object({
-    approvedPersonIds: z.array(z.uuidv4()).default([]),
-    excludeFaceIds: z.array(z.uuidv4()).optional(),
-    manualMove: FaceRepairManualMoveSchema.optional(),
-  })
-  .refine((value) => value.approvedPersonIds.length > 0 || value.manualMove !== undefined, {
-    error: 'approvedPersonIds must be non-empty unless manualMove is provided',
-    path: ['approvedPersonIds'],
-  })
-  .meta({ id: 'FaceRepairApplyRequestDto' });
-export class FaceRepairApplyRequestDto extends createZodDto(FaceRepairApplyRequestSchema) {}
-
-export const FaceRepairApplyResponseSchema = z
-  .object({ moved: z.number(), skipped: z.number() })
-  .meta({ id: 'FaceRepairApplyResponseDto' });
-export class FaceRepairApplyResponseDto extends createZodDto(FaceRepairApplyResponseSchema) {}
-
 const FlaggedFaceSchema = z.object({ assetFaceId: z.string(), suspectedOwnerId: z.string() });
 export const FaceRepairPersonFacesSchema = z
   .object({ personId: z.string(), flaggedFaces: z.array(FlaggedFaceSchema) })
